algorithm/smart-attack/requirements.txt
pycryptodome == 3.17
